refactor: naming + files after merge
