+++ frontend/src/views/AddFuelCode.jsx
@@ -1,0 +1,3 @@
+export const AddFuelCode = () => {
+  return <div>AddFuelCode</div>
+}
