+++ frontend/src/views/AddOrganization/index.jsx
@@ -1,0 +1,1 @@
+export { AddOrganization } from './AddOrganization'
