+++ frontend/src/views/AddTransaction.jsx
@@ -1,0 +1,3 @@
+export const AddTransaction = () => {
+  return <div>AddTransaction</div>
+}
