--- frontend/src/views/NewFuelCode.jsx
@@ -1,3 +1,0 @@
-export const NewFuelCode = () => {
-  return <div>NewFuelCode</div>
-}
