--- frontend/src/views/NewOrg.jsx
@@ -1,3 +1,0 @@
-export const NewOrg = () => {
-  return <div>NewOrg</div>
-}
