--- frontend/src/views/NewTransaction.jsx
@@ -1,3 +1,0 @@
-export const NewTransaction = () => {
-  return <div>NewTransaction</div>
-}
